kademlia: Workaround for dealing with not implemented `PUT_VALUE` ACKs (#430)
Because previously litep2p didn't send/receive `PUT_VALUE` ACKs, we
can't just enable this mechanism — this would lead to failures when
older remote peers don't behave up to spec. This PR deals with this
situation by implementing a workaround to handle such peers:

1. Track `PUT_VALUE` send successes instead of ACK responses to
determine the query success.
2. Ignore send errors when the remote doesn't read the `PUT_VALUE` ACK
message.

This should be seen as a temporary measure until most of the networks
updates to litep2p with `PUT_VALUE` ACK support (this PR or newer).
Suggested timing is in a year after releasing the version with this PR.
This is tracked by issue
https://github.com/paritytech/litep2p/issues/429.

Builds upon https://github.com/paritytech/litep2p/pull/427.