fix: multistream-select negotiation on outbound substream over webrtc (#465)
This PR changes how multistream-select messages received on an opening
outbound substream are handled on webrtc connections. For details please
see the issue description.

The PR builds on top of #441 for the str0m update. But looking at it
again I realized that it also made changes to multistream-select
handling. I'll look into what those are and how they relate to my
changes. In any case, I believe my changes bring the implementation
closer to compliance with the multistream-select spec and enable
interoperability with smoldot.

fixes #464

---------

Co-authored-by: Timothy Wu <tim.wu@chainsafe.io>
Co-authored-by: Dmitry Markin <dmitry@markin.tech>