tcp: Exit connections when events fail to propagate to protocols (#506)
This PR ensures that TCP connections exit when they cannot propagate
events to the installed protocols.

This may happen when the installed protocol panics or exits abruptly.
Therefore, the protocol receiver is dropped, and the connection serves
no purpose.

This PR aligns the TCP implementation with the Webscoket one.

Detected by:
- https://github.com/paritytech/polkadot-sdk/issues/10260

Signed-off-by: Alexandru Vasile <alexandru.vasile@parity.io>